+++ macos_pkg_builder/__init__.py
@@ -18,8 +18,8 @@ Library usage:
     >>> pkg_obj.build()
 """
 
-__title__:        str = "nekrosis"
-__version__:      str = "1.1.0"
+__title__:        str = "macos_pkg_builder"
+__version__:      str = "1.1.1"
 __description__:  str = "macOS Package Builder Library, wrapping the `pkgbuild` and `productbuild` tools."
 __url__:          str = "https://github.com/ripeda/macOS-Pkg-Builder"
 __author__:       str = "RIPEDA Consulting"